+++ test/spec/unit/DeferredUnitSpec.js
@@ -72,11 +72,10 @@ describe("Unit tests for Deferred Objects", function() {
 
   });
 
-  it("should cleanup mock ajax methods ", function() {
-
+  it("should clear ajax mocks", function() {
     runs(function() {
-        // Cleanup ajax mock
-        clearAllAjaxMocks();
+      // Cleanup ajax mock
+      clearAllAjaxMocks();
     });
   });