+++ test/spec/unit/UnitTestHelper.js
@@ -10,8 +10,25 @@ function expectEndpoint(obj, methodParams, sdkMethod, verb, endPoint) {
     }
   });
   var calledBack = false;
-  obj[sdkMethod].apply(obj, methodParams);
-  expect(calledBack).toEqual(true);
+
+  runs(function() {
+    methodParams = methodParams || [];
+    methodParams.push({
+      inspectParams: function(model, params, method) {
+        calledBack = true;
+        expect(params).toCall(endPoint, verb);
+      }
+    });
+    obj[sdkMethod].apply(obj, methodParams);
+  });
+
+  waitsFor(function() {
+    return calledBack === true;
+  });
+
+  runs(function() {
+    expect(calledBack).toEqual(true);
+  });
 }
 
 beforeEach(function() {
@@ -30,6 +47,14 @@ beforeEach(function() {
       } catch(e) {
         return false;
       }
+    },
+
+    toStartWith: function(value) {
+      return this.actual.indexOf(value) === 0;
+    },
+
+    toEndWith: function(value) {
+      return this.actual.indexOf(value) === this.length - value.length;
     }
   });